+++ packages/dynamic-flows/src/flow/client/HttpClient.js
@@ -1,14 +1,14 @@
-class HttpClient {
-  init(params = {}) {
+export class HttpClient {
+  constructor(params = {}) {
     this.baseUrl = params.baseUrl;
-
-    return this;
+    this.headers = params.headers || {};
   }
 
   async request({ action, data }) {
     const { url, method } = action;
 
-    const endpoint = `${this.baseUrl}${url}`;
+    const endpoint = startsWithHTTP(url) ? url : `${this.baseUrl}${url}`;
+
     const body = method === 'GET' ? undefined : JSON.stringify(data);
 
     return fetch(endpoint, {
@@ -16,6 +16,7 @@ class HttpClient {
       headers: {
         'Content-Type': 'application/json',
         'X-Access-Token': 'Tr4n5f3rw153',
+        ...this.headers,
       },
       credentials: 'include',
       body,
@@ -28,4 +29,8 @@ class HttpClient {
   }
 }
 
-export const httpClient = new HttpClient();
+function startsWithHTTP(url = '') {
+  return ['https://', 'http://'].some(
+    (prefix) => url.slice(0, prefix.length) === prefix && url.length > prefix.length,
+  );
+}