+++ packages/dynamic-flows/src/index.js
@@ -1,5 +1,5 @@
 export { convertStepToLayout, inlineReferences } from './step/layoutStep/layoutService';
-export { default as DynamicFlow } from './flow';
+export { default as DynamicFlow, HttpClient } from './flow';
 export { default as Field } from './field';
 export { default as FormControl } from './formControl';
 export { default as JsonSchemaForm } from './jsonSchemaForm';