+++ packages/dynamic-flows/src/layout/image/index.js
@@ -1,6 +1,8 @@
 import { Image } from '@transferwise/components';
 import Types from 'prop-types';
+import { useEffect, useState } from 'react';
 
+import { useBaseUrl } from '../../common/contexts/baseUrlContext/BaseUrlContext';
 import { marginModel } from '../models';
 import { getMarginBottom } from '../utils';
 
@@ -25,25 +27,70 @@ const getImageClasses = (image) => {
   return `img-responsive ${margin}`;
 };
 
+const readImageBlobAsDataURL = (imageBlob) =>
+  new Promise((resolve, reject) => {
+    const reader = new FileReader();
+
+    reader.addEventListener('loadend', () => resolve(reader.result));
+    reader.addEventListener('error', (error) => reject(error));
+
+    reader.readAsDataURL(imageBlob);
+  });
+
+const getImageSource = async (baseUrl, image) => {
+  try {
+    const url = new URL(image.url, window?.location?.origin);
+    if (image.url?.indexOf(url.pathname) === 0) {
+      url.pathname = baseUrl + url.pathname;
+    }
+
+    if (url.origin !== window?.location?.origin) {
+      return image.url;
+    }
+
+    return fetch(url, {
+      method: 'GET',
+      headers: {
+        'Content-Type': 'image/*',
+        'X-Access-Token': 'Tr4n5f3rw153',
+      },
+      credentials: 'same-origin',
+    })
+      .then((response) => response.blob())
+      .then(readImageBlobAsDataURL);
+  } catch {
+    return image.url;
+  }
+};
+
 const DynamicImage = (props) => {
   const image = props.component;
 
+  const baseUrl = useBaseUrl();
+
+  const [imageSource, setImageSource] = useState('');
+
+  useEffect(() => {
+    getImageSource(baseUrl, image).then(setImageSource);
+  }, [image.url]);
+
   const imageProps = {
     alt: image.text,
-    src: image.url,
+    src: imageSource,
     stretch: true,
     shrink: true,
   };
 
+  const getImage = () =>
+    imageSource ? <Image className={getImageClasses(image)} {...imageProps} /> : null;
+
   if (!image.size || image.size === 'xl') {
-    return <Image className={getImageClasses(image)} {...imageProps} />;
+    return getImage();
   }
 
   return (
     <div className="row m-b-0">
-      <div className={getWrapperClasses(image)}>
-        <Image className={getImageClasses(image)} {...imageProps} />
-      </div>
+      <div className={getWrapperClasses(image)}>{getImage()}</div>
     </div>
   );
 };