+++ packages/dynamic-flows/src/ssr.spec.js
@@ -6,37 +6,46 @@ import { renderToString } from 'react-dom/server';
 
 import * as components from '.';
 
-const excluded = new Set(['useSnackbar', 'convertStepToLayout', 'inlineReferences', 'DynamicFlow']);
+const excluded = new Set([
+  'useSnackbar',
+  'convertStepToLayout',
+  'inlineReferences',
+  'DynamicFlow',
+  'HttpClient',
+]);
 
 function isNotExcluded(componentName) {
   return !excluded.has(componentName);
 }
 
+const componentNames = Object.keys(components).filter(isNotExcluded);
+
 describe('Server side rendering', () => {
-  it('can render components', () => {
-    const componentNames = Object.keys(components).filter(isNotExcluded);
+  it('expects some components', () => {
     expect(componentNames.length).toBeGreaterThan(0);
+  });
 
-    // stick all possible properties we might need to render all components in here
-    const allProps = {
-      schema: {},
-      components: [],
-      onAction: jest.fn(),
-      onModelChange: jest.fn(),
-      onChange: jest.fn(),
-      submitted: false,
-      name: 'name',
-      field: {
-        control: 'text',
-        type: 'string',
-        label: 'hello',
-      },
-      onPersistAsync: jest.fn(),
-      baseUrl: 'some-url',
-    };
-    const overrideProps = {};
+  componentNames.forEach((componentName) => {
+    it(`can render ${componentName}`, () => {
+      // stick all possible properties we might need to render all components in here
+      const allProps = {
+        schema: {},
+        components: [],
+        onAction: jest.fn(),
+        onModelChange: jest.fn(),
+        onChange: jest.fn(),
+        submitted: false,
+        name: 'name',
+        field: {
+          control: 'text',
+          type: 'string',
+          label: 'hello',
+        },
+        onPersistAsync: jest.fn(),
+        baseUrl: 'some-url',
+      };
+      const overrideProps = {};
 
-    componentNames.forEach((componentName) => {
       const Component = components[componentName];
       const newProps = { ...allProps };